pyproject.toml
[tool.poetry]
name = "tech-assignment"
version = "0.1.0"
description = ""
authors = ["Mark B. <mark.bobkov.13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
python-dotenv = "^1.0.1"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.31"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.1"}
drf-standardized-errors = {extras = ["openapi"], version = "^0.12.6"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
pytest-django = "^4.8.0"
pytest-drf = "^1.1.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
